refactor(db): relocate persist mapping into db/ and symmetrize writeCollection
Move src/shell/commands/snapshot/generate/storage.ts into src/db/persist.ts as the write counterpart to rebuild.ts; db/ now owns both halves of snapshot persistence. Symmetrize the writeCollection group so all four writers take their own unit-type array (KuralFile[]/KuralType[]/KuralFunction[]/KuralDirectory[]) and writeUnits handles the extraction up front, eliminating writeDirectories' signature asymmetry.
